Cleanup of scripts and empty directories
And some Rider-suggested optimisations
