--- Assets/Prefabs/Alwin test.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff908703d041346ea93d9004ecebb7bf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
