--- Assets/Scripts/Camera.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5de65097b1a94b87a1af2d63ec16e4d1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
