--- Assets/Scripts/Menu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a423879db798a48fa91bc95f10ba685f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
