+++ Assets/Scripts/Player/Jump.cs
@@ -1,42 +1,44 @@
-using System.Collections;
+using IK;
 using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 
-public class Jump : MonoBehaviour
+namespace Player
 {
-    [SerializeField] private List<IKFootHandler> feet;
-    [SerializeField] private float jumpHeight;
-    private Rigidbody _rb;
-    public bool canJump;
-    void Start()
-    {
-        _rb = GetComponent<Rigidbody>();
-    }
+	public class Jump : MonoBehaviour
+	{
+		[SerializeField] private List<IKFootHandler> feet;
+		[SerializeField] private float jumpHeight;
+		private Rigidbody _rb;
+		public bool canJump;
 
-    // Update is called once per frame
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Space) && canJump)
-        {
-            foreach (IKFootHandler foot in feet)
-            {
-                foot.enabled = false;
-            }
-            DoJump();
-        }
-    }
+		private void Start()
+		{
+			_rb = GetComponent<Rigidbody>();
+		}
 
-    private void DoJump()
-    {
-        _rb.AddForce(Vector3.up * jumpHeight * Time.deltaTime, ForceMode.Impulse);
+		private void Update()
+		{
+			if (Input.GetKeyDown(KeyCode.Space) && canJump)
+			{
+				foreach(IKFootHandler foot in feet)
+				{
+					foot.enabled = false;
+				}
+				DoJump();
+			}
+		}
 
-        _rb.mass = 10f;
-        _rb.drag = .5f;
-        canJump = false;
-        foreach (IKFootHandler foot in feet)
-        {
-            foot.enabled = true;
-        }
-    }
+		private void DoJump()
+		{
+			_rb.AddForce(Vector3.up * (jumpHeight * Time.deltaTime), ForceMode.Impulse);
+
+			_rb.mass = 10f;
+			_rb.drag = .5f;
+			canJump = false;
+			foreach(IKFootHandler foot in feet)
+			{
+				foot.enabled = true;
+			}
+		}
+	}
 }
